# public/openapi.yaml
openapi: 3.0.3
info:
  title: GreenQuote API
  version: 1.0.0
  description: |
    Solar financing pre-qualification API.
    Authenticated routes use an httpOnly session cookie (`gq_session`) set by login/register.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in; session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gq_session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
...
    UserRole:
      type: string
      enum: [user, admin]
...